assistant/requirements.txt
python-dotenv 
langchain 
langchain-openai 
langchain-pinecone
langchain-community 
pinecone-client
streamlit
wikipedia
GitPython
unstructured